Refactor columns and insertion
When a directory is selected in the view, it should pass it's name column down to some project adding activity so that we can define which directory the project is in.

When clicking a project, it should also do the same thing to the note adding activity.

There's a lot of speghetti code here. Needs major refactoring.