Import Image update
Can import an image from external storage into the app's internal storage. The image and its path is saved in the database.

How to procede with this feature, add a way for a user to title the image with a unique name. Currently using some dummy title that will create/overwrite a file of that name.